move Logger interface to services
